+++ crates/ruma-state-res/benches/state_res_bench.rs
@@ -79,6 +79,7 @@ fn resolution_shallow_auth_chain(c: &mut Criterion) {
                     })
                     .collect(),
                 |id| ev_map.get(id).map(Arc::clone),
+                |_| unreachable!(),
             ) {
                 Ok(state) => state,
                 Err(e) => panic!("{e}"),
@@ -138,6 +139,7 @@ fn resolve_deeper_event_set(c: &mut Criterion) {
                     })
                     .collect(),
                 |id| inner.get(id).map(Arc::clone),
+                |_| unreachable!(),
             )
             .unwrap_or_else(|_| panic!("resolution failed during benchmarking"));
         });